pyproject.toml
[build-system]
requires = ["setuptools", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "MailApp"
version = "0.1.0"
description = "A simple mail application"
authors = [
    { name = "NKNS", email = "hcaxkcab@gmail.com" }
]
dependencies = [
    "requests>=2.19.1,<2.30",
    "flask",
    "certifi==2024.2.2",
    "cffi==1.16.0",
    "charset-normalizer==3.3.2",
    "cryptography==42.0.2",
    "gcloud==0.18.3",
    "google-auth==2.27.0",
    "google-auth-httplib2==0.2.0",
    "googleapis-common-protos==1.62.0",
    "httplib2==0.22.0",
    "idna==3.6",
    "jwcrypto==1.5.1",
    "oauth2client==4.1.3",
    "protobuf==4.25.2",
    "pyasn1==0.5.1",
    "pyasn1-modules==0.3.0",
    "pycparser==2.21",
    "PyJWT==2.8.0",
    "PyQt6==6.8.1",
    "PyQt6_sip>=13.8,<14",
    "Pyrebase4==4.7.1",
    "python-dotenv==1.0.1",
    "python-jwt==4.1.0",
    "requests-toolbelt>=0.7.1,<1.0",
    "rsa==4.9",
    "six==1.16.0",
    "urllib3>=1.21.1,<2.0"
]